requirements.txt
beautifulsoup4==4.8.1
python-dotenv==0.10.3
soupsieve==1.9.5
terminaltables==3.1.0
